requirements.txt
requests
python-dotenv
openai
pandas
openpyxl
